+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 
 New functionality added in a backwards-compatible manner:
 * [OLMIS-7307](https://openlmis.atlassian.net/browse/OLMIS-7307): Added extension module for postprocess create and update products.
+* [OLMIS-7304](https://openlmis.atlassian.net/browse/OLMIS-7304): Added sending orderable object to one-network-integration-service.